GCW-3503 Create useSafeSetState hook
