+++ src/hooks/useSafeSetState.ts
@@ -1,0 +1,18 @@
+import { useEffect, useRef } from "react";
+
+export const useSafeSetState = () => {
+  const isMounted = useRef(false);
+
+  const safeSetState = (setStateCallback: () => React.SetStateAction<any>) => {
+    if (isMounted.current) setStateCallback();
+  };
+
+  useEffect(() => {
+    isMounted.current = true;
+    return () => {
+      isMounted.current = false;
+    };
+  }, []);
+
+  return safeSetState;
+};
